pyproject.toml
[project]
name = "careerdevelopmenttool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
